requirements.txt
pandas
praw
tweepy
python-dateutil